Parse log level
